Make QueryBuilderProtoConverter and QueryBuilderProtoConverterRegistry an SPI
