+++ modules/transport-grpc/spi/build.gradle
@@ -1,0 +1,23 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * The OpenSearch Contributors require contributions made to
+ * this file be licensed under the Apache-2.0 license or a
+ * compatible open source license.
+ */
+
+apply plugin: 'opensearch.build'
+apply plugin: 'opensearch.publish'
+
+base {
+  group = 'org.opensearch.plugin'
+  archivesName = 'transport-grpc-spi'
+}
+
+dependencies {
+  api project(":server")
+  api "org.opensearch:protobufs:0.6.0"
+}
+
+// no tests...yet?
+test.enabled = false
